+++ internal/bson/document.go
@@ -179,12 +179,16 @@ func (doc *Document) Command() string {
 	return doc.fields[0].name
 }
 
-// Encode encodes BSON document.
+// Encode encodes non-nil BSON document.
 //
 // TODO https://github.com/FerretDB/FerretDB/issues/3759
 // This method should accept a slice of bytes, not return it.
 // That would allow to avoid unnecessary allocations.
+//
+// Receiver must not be nil.
 func (doc *Document) Encode() (RawDocument, error) {
+	must.NotBeZero(doc)
+
 	size := sizeAny(doc)
 	buf := bytes.NewBuffer(make([]byte, 0, size))
