+++ internal/bson/raw_array.go
@@ -19,6 +19,7 @@ import (
 	"strconv"
 
 	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
+	"github.com/FerretDB/FerretDB/internal/util/must"
 )
 
 // RawArray represents a single BSON array in the binary encoded form.
@@ -27,16 +28,23 @@ import (
 type RawArray []byte
 
 // Encode returns itself to implement the [AnyArray] interface.
+//
+// Receiver must not be nil.
 func (raw RawArray) Encode() (RawArray, error) {
+	must.BeTrue(raw != nil)
 	return raw, nil
 }
 
-// Decode decodes a single BSON array that takes the whole byte slice.
+// Decode decodes a single BSON array that takes the whole not-nil byte slice.
 //
 // Only top-level elements are decoded;
 // nested documents and arrays are converted to RawDocument and RawArray respectively,
 // using raw's subslices without copying.
+//
+// Receiver must not be nil.
 func (raw RawArray) Decode() (*Array, error) {
+	must.BeTrue(raw != nil)
+
 	res, err := raw.decode(decodeShallow)
 	if err != nil {
 		return nil, lazyerrors.Error(err)
@@ -45,10 +53,14 @@ func (raw RawArray) Decode() (*Array, error) {
 	return res, nil
 }
 
-// DecodeDeep decodes a single valid BSON array that takes the whole byte slice.
+// DecodeDeep decodes a single valid BSON array that takes the whole not-nil byte slice.
 //
 // All nested documents and arrays are decoded recursively.
+//
+// Receiver must not be nil.
 func (raw RawArray) DecodeDeep() (*Array, error) {
+	must.BeTrue(raw != nil)
+
 	res, err := raw.decode(decodeDeep)
 	if err != nil {
 		return nil, lazyerrors.Error(err)
