+++ internal/bson/raw_document.go
@@ -27,16 +27,23 @@ import (
 type RawDocument []byte
 
 // Encode returns itself to implement the [AnyDocument] interface.
+//
+// Receiver must not be nil.
 func (raw RawDocument) Encode() (RawDocument, error) {
+	must.BeTrue(raw != nil)
 	return raw, nil
 }
 
-// Decode decodes a single BSON document that takes the whole byte slice.
+// Decode decodes a single BSON document that takes the whole not-nil byte slice.
 //
 // Only top-level fields are decoded;
 // nested documents and arrays are converted to RawDocument and RawArray respectively,
 // using raw's subslices without copying.
+//
+// Receiver must not be nil.
 func (raw RawDocument) Decode() (*Document, error) {
+	must.BeTrue(raw != nil)
+
 	res, err := raw.decode(decodeShallow)
 	if err != nil {
 		return nil, lazyerrors.Error(err)
@@ -45,10 +52,14 @@ func (raw RawDocument) Decode() (*Document, error) {
 	return res, nil
 }
 
-// DecodeDeep decodes a single valid BSON document that takes the whole byte slice.
+// DecodeDeep decodes a single valid BSON document that takes the whole not-nil byte slice.
 //
 // All nested documents and arrays are decoded recursively.
+//
+// Receiver must not be nil.
 func (raw RawDocument) DecodeDeep() (*Document, error) {
+	must.BeTrue(raw != nil)
+
 	res, err := raw.decode(decodeDeep)
 	if err != nil {
 		return nil, lazyerrors.Error(err)
